requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.0.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dataclasses-json==0.5.14
docarray==0.32.1
entrypoints==0.4
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
hnswlib==0.7.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.265
langsmith==0.0.22
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.4
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pypdf==3.15.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
semver==3.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
streamlit==1.25.0
streamlit-chat==0.1.1
tenacity==8.2.3
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
watchdog==3.0.0
yarl==1.9.2
zipp==3.16.2
